requirements.txt
Click==7.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
PyJWT==1.7.1
python-dotenv==0.10.3
SQLAlchemy==1.3.8
Werkzeug==0.15.6
